+++ order-service/src/main/java/com/ecommerce/orderservice/application/dto/UpdateOrderRequest.java
@@ -1,4 +1,4 @@
-package com.ecommerce.orderservice.application.request;
+package com.ecommerce.orderservice.application.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
