+++ order-service/src/main/java/com/ecommerce/orderservice/infrastructure/adapter/feign/PaymentFeignAdapter.java
@@ -3,8 +3,8 @@ package com.ecommerce.orderservice.infrastructure.adapter.feign;
 import com.ecommerce.orderservice.application.dto.PaymentAuthorizationResponse;
 import com.ecommerce.orderservice.application.dto.RefundResponse;
 import com.ecommerce.orderservice.application.port.out.PaymentServicePort;
-import com.ecommerce.orderservice.application.request.PaymentAuthorizationRequest;
-import com.ecommerce.orderservice.application.request.RefundRequest;
+import com.ecommerce.orderservice.application.dto.PaymentAuthorizationRequest;
+import com.ecommerce.orderservice.application.dto.RefundRequest;
 import com.ecommerce.orderservice.domain.exception.OrderValidationException;
 import com.ecommerce.shared.exception.ExceptionError;
 import com.ecommerce.shared.exception.ResourceNotFoundException;
