+++ order-service/src/main/java/com/ecommerce/orderservice/infrastructure/adapter/feign/ProductFeignAdapter.java
@@ -3,7 +3,7 @@ package com.ecommerce.orderservice.infrastructure.adapter.feign;
 import com.ecommerce.orderservice.application.dto.ProductAvailabilityResponse;
 import com.ecommerce.orderservice.application.dto.ProductResponse;
 import com.ecommerce.orderservice.application.port.out.ProductServicePort;
-import com.ecommerce.orderservice.application.request.OrderItemRequest;
+import com.ecommerce.orderservice.application.dto.OrderItemRequest;
 import com.ecommerce.orderservice.domain.exception.OrderValidationException;
 import com.ecommerce.shared.exception.ExceptionError;
 import com.ecommerce.shared.exception.ResourceNotFoundException;
