+++ order-service/src/main/java/com/ecommerce/orderservice/infrastructure/adapter/feign/ProductFeignClient.java
@@ -2,7 +2,7 @@ package com.ecommerce.orderservice.infrastructure.adapter.feign;
 
 import com.ecommerce.orderservice.application.dto.ProductAvailabilityResponse;
 import com.ecommerce.orderservice.application.dto.ProductResponse;
-import com.ecommerce.orderservice.application.request.OrderItemRequest;
+import com.ecommerce.orderservice.application.dto.OrderItemRequest;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.*;
